+++ MeetingCoach/MeetingCoach/Views/SimulationTimelineView.swift
@@ -271,6 +271,7 @@ struct SimNudgeCardView: View {
     }
 
     private var urgencyColor: Color {
+        if nudge.type.isPositive { return .green }
         switch nudge.urgency {
         case .low: return .gray
         case .med: return .blue